requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
python-multipart
opencv-python
easyocr
numpy
Pillow